requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-trade-api==3.2.0
annotated-types==0.6.0
anthropic==0.45.2
anyio==4.3.0
attrs==23.2.0
Automat==22.10.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.9
constantly==23.10.4
contourpy==1.2.1
courlan==1.0.0
coverage==7.6.7
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dateparser==1.2.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
fastapi==0.115.7
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.159.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.67.1
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.8.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.0
jusText==3.0.0
kiwisolver==1.4.5
langcodes==3.3.0
lxml==5.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mem0ai==0.1.47
monotonic==1.6
mplfinance==0.12.10b0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.0.5
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.60.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parsel==1.9.1
pdfkit==1.0.0
pillow==10.3.0
portalocker==2.10.1
posthog==3.9.2
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyDispatcher==2.0.7
pygame==2.5.2
pymongo==4.6.3
PyMuPDF==1.25.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytz==2024.1
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.1
qdrant-client==1.13.2
queuelib==1.6.2
regex==2023.12.25
renko==0.0.16
renkodf==1.1
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
Scrapy==2.11.1
sentence-transformers==2.6.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.37
starlette==0.45.3
stripe==11.2.0
sympy==1.12
ta==0.11.0
threadpoolctl==3.4.0
tld==0.13
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
trafilatura==1.8.1
transformers==4.39.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.34.0
w3lib==2.1.2
websocket-client==1.7.0
websockets==10.4
Werkzeug==3.0.2
yarl==1.9.4
zope.interface==6.3
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
google-cloud-aiplatform 
ipython
torch
datasets
sentence-transformers
scikit-learn
